chore: extract fetch_abi_from_etherscan and use &Config (#11426)
