+++ crates/cast/src/cmd/artifact.rs
@@ -1,6 +1,6 @@
 use super::{
     creation_code::{fetch_creation_code_from_etherscan, parse_code_output},
-    interface::{fetch_abi_from_etherscan, load_abi_from_file},
+    interface::load_abi_from_file,
 };
 use alloy_primitives::Address;
 use alloy_provider::Provider;
@@ -10,9 +10,11 @@ use foundry_cli::{
     opts::{EtherscanOpts, RpcOpts},
     utils::{self, LoadConfig},
 };
-use foundry_common::fs;
+use foundry_common::{abi::fetch_abi_from_etherscan, fs};
 use serde_json::json;
 use std::path::PathBuf;
+
+foundry_config::impl_figment_convert!(ArtifactArgs, etherscan, rpc);
 
 /// CLI arguments for `cast artifact`.
 #[derive(Parser)]
@@ -45,24 +47,25 @@ pub struct ArtifactArgs {
 
 impl ArtifactArgs {
     pub async fn run(self) -> Result<()> {
-        let Self { contract, mut etherscan, rpc, output: output_location, abi_path } = self;
+        let mut config = self.load_config()?;
 
-        let config = rpc.load_config()?;
+        let Self { contract, output: output_location, abi_path, etherscan: _, rpc: _ } = self;
+
         let provider = utils::get_provider(&config)?;
         let chain = provider.get_chain_id().await?;
-        etherscan.chain = Some(chain.into());
+        config.chain = Some(chain.into());
 
         let abi = if let Some(ref abi_path) = abi_path {
             load_abi_from_file(abi_path, None)?
         } else {
-            fetch_abi_from_etherscan(contract, &etherscan).await?
+            fetch_abi_from_etherscan(contract, &config).await?
         };
 
         let (abi, _) = abi.first().ok_or_else(|| eyre::eyre!("No ABI found"))?;
 
-        let bytecode = fetch_creation_code_from_etherscan(contract, &etherscan, provider).await?;
+        let bytecode = fetch_creation_code_from_etherscan(contract, &config, provider).await?;
         let bytecode =
-            parse_code_output(bytecode, contract, &etherscan, abi_path.as_deref(), true, false)
+            parse_code_output(bytecode, contract, &config, abi_path.as_deref(), true, false)
                 .await?;
 
         let artifact = json!({
