+++ crates/cast/src/cmd/interface.rs
@@ -2,10 +2,10 @@ use alloy_json_abi::{ContractObject, JsonAbi};
 use alloy_primitives::Address;
 use clap::Parser;
 use eyre::{Context, Result};
-use foundry_block_explorers::Client;
 use foundry_cli::{opts::EtherscanOpts, utils::LoadConfig};
 use foundry_common::{
     ContractsByArtifact,
+    abi::fetch_abi_from_etherscan,
     compile::{PathOrContractInfo, ProjectCompiler},
     find_target_path, fs, shell,
 };
@@ -65,7 +65,7 @@ impl InterfaceArgs {
             load_abi_from_file(&contract, name)?
         } else {
             match Address::from_str(&contract) {
-                Ok(address) => fetch_abi_from_etherscan(address, &etherscan).await?,
+                Ok(address) => fetch_abi_from_etherscan(address, &etherscan.load_config()?).await?,
                 Err(_) => load_abi_from_artifact(&contract)?,
             }
         };
@@ -135,20 +135,6 @@ fn load_abi_from_artifact(path_or_contract: &str) -> Result<Vec<(JsonAbi, String
         maybe_abi.as_ref().ok_or_else(|| eyre::eyre!("Failed to fetch lossless ABI"))?;
 
     Ok(vec![(abi.clone(), contract.name().unwrap_or(name).to_string())])
-}
-
-/// Fetches the ABI of a contract from Etherscan.
-pub async fn fetch_abi_from_etherscan(
-    address: Address,
-    etherscan: &EtherscanOpts,
-) -> Result<Vec<(JsonAbi, String)>> {
-    let config = etherscan.load_config()?;
-    let chain = config.chain.unwrap_or_default();
-    let api_version = config.get_etherscan_api_version(Some(chain));
-    let api_key = config.get_etherscan_api_key(Some(chain)).unwrap_or_default();
-    let client = Client::new_with_api_version(chain, api_key, api_version)?;
-    let source = client.contract_source_code(address).await?;
-    source.items.into_iter().map(|item| Ok((item.abi()?, item.contract_name))).collect()
 }
 
 /// Converts a vector of tuples containing the ABI and contract name into a vector of
