+++ crates/common/src/abi.rs
@@ -1,7 +1,7 @@
 //! ABI related helper functions.
 
 use alloy_dyn_abi::{DynSolType, DynSolValue, FunctionExt, JsonAbiExt};
-use alloy_json_abi::{Error, Event, Function, Param};
+use alloy_json_abi::{Error, Event, Function, JsonAbi, Param};
 use alloy_primitives::{Address, LogData, hex};
 use eyre::{Context, ContextCompat, Result};
 use foundry_block_explorers::{
@@ -135,6 +135,19 @@ pub fn get_indexed_event(mut event: Event, raw_log: &LogData) -> Event {
         })
     }
     event
+}
+
+/// Fetches the ABI of a contract from Etherscan.
+pub async fn fetch_abi_from_etherscan(
+    address: Address,
+    config: &foundry_config::Config,
+) -> Result<Vec<(JsonAbi, String)>> {
+    let chain = config.chain.unwrap_or_default();
+    let api_version = config.get_etherscan_api_version(Some(chain));
+    let api_key = config.get_etherscan_api_key(Some(chain)).unwrap_or_default();
+    let client = Client::new_with_api_version(chain, api_key, api_version)?;
+    let source = client.contract_source_code(address).await?;
+    source.items.into_iter().map(|item| Ok((item.abi()?, item.contract_name))).collect()
 }
 
 /// Given a function name, address, and args, tries to parse it as a `Function` by fetching the
